+++ docs/source/tutorials/getting_started/data_query_2.py
@@ -130,5 +130,4 @@ pass
 # |
 # |
 #
-# Download this Tutorial
-# ======================
+# .. rubric:: Download this Tutorial