+++ docs/source/tutorials/getting_started/data_saving.py
@@ -67,5 +67,4 @@ pass
 # |
 # |
 #
-# Download this Tutorial
-# ======================
+# .. rubric:: Download this Tutorial