+++ docs/source/tutorials/images/image_plotting.py
@@ -109,5 +109,4 @@ figure
 # |
 # |
 #
-# Download this Tutorial
-# ----------------------
+# .. rubric:: Download this Tutorial