+++ docs/source/tutorials/images/image_plotting_2.py
@@ -35,5 +35,4 @@ figure
 # |
 # |
 #
-# Download this Tutorial
-# ----------------------
+# .. rubric:: Download this Tutorial