+++ docs/source/tutorials/images/image_query.py
@@ -225,5 +225,4 @@ figure
 # |
 # |
 #
-# Download this Tutorial
-# ----------------------
+# .. rubric:: Download this Tutorial