+++ src/ATK/io/plot_io.py
@@ -19,7 +19,9 @@ from ..structures.Target import Target
 from ..utilities.mapping import build_map
 
 
-def dispatch_plotting(all_figures: list, plotting_func: FunctionType, structure: DataSet, containers: list[Container], target: Target | None = None, **kwargs):
+def dispatch_plotting(
+    all_figures: list, plotting_func: FunctionType, structure: DataSet, containers: list[Container], target: Target | None = None, **kwargs
+):
     # plot .data containers individually (e.g. images)
     if structure._plot_method == "individual":
         figures = [plotting_func(ctnr, **kwargs) for ctnr in containers]